Issue #50: Add a Terminology Section
This is a first pass, placing the Terminology section just before
Acknowledgements.

I made a pass through, attempting to ensure the first use of a term within
a given section was always linked back to the definition.

I also made an effort to always spell out "WebAuthn Relying Party" for the first
use in each section, and remove the use of "RP" (excepting the use of "RP ID").

There are some more whitespace cleanups as well, though I tried to minimize them.