Rename imageUri to imageURL
Also removed the level from the metadata since it was appending a 1 to
the document title. Fixes #59. I decided not to mandate manifest
formatting of the icon to avoid the added complexity. Authenticators
would have to implement it, and we'd have to explore all the things a
bad person could do (e.g. have very different looking avatars at
different resolutions).