Markdown cleanup (#75)
* Markdown cleanup

No changes to substance here. Goal of this change is to make the spec
easier to work on, not to fix the content.

- Wrap all lines at 128 chars (wide enough for modern screens, small
enough to fit in most commonly-sized windows)
- Remove extraneous HTML tags
- Convert HTML tags to Markdown equivalents
- Define macros for a couple of commonly used phrases so it's easier to
type and to keep consistency throughout the spec

* Fixed links

Bikeshed does not support markdown for links - I had realized this
earlier but had not reverted all the links to href form. Finishing the
job.