Simplify attestation structure by moving details into authnr model
Moved most of the attestation details into the authenticator model
section, and flattened the structures for simplicity. Updated spec
language to match.

It may be possible to further simplify the attestation details in the
authenticator model section, but I would prefer that an attestation
expert weigh in on that.

Also a few other small formatting fixups.