Clarify security model and remove vestiges of "native API" language
Fixes #58 by adding verbiage in the API section about the two layers in
the WebAuthn security model. This discussion is ongoing with the TAG, so
we will open a new issue if there are follow-ups.

Fixes #20, #21, #23 by removing mentions of native apps and APIs, as
these are out of scope for this specification.