Remove ScopedCredentialParameters tuple (by unpairing type & algorithm)
The procedure for normalizing algorithms and types doesn't actually use the
types. Keeping these together in an object gives a kind of flexibility, but
after consideration, I don't see much purpose in defining different algorithms
for specifically different authenticator types.

This changes the makeCredential() method to instead have a list of supported
types and a list of supported algorithms, and perform its procedure that way.