API aesthetic updates (#142)
* Change the parent object from "window" to "navigator"

The HTML5 `window` object represents an open window in the browser, whereas
the `navigator` object represents the browser itself. Since authenticators
belong to the whole system, not one window, the object model makes more sense
to be a property attached to `navigator`.

* Rename `navigator.webauthn` to `navigator.authentication`.

Generally it's frowned upon to name a Web API "web-". I considered using
"authn", but I feel the verbosity here is justified. Most users will assign
the object to a local variable for ease of use anyway, just as our example
code does.

* Use CryptoKey for ScopedCredentialInfo.publicKey

* Rename example extension IDs and provide guidance about dot-notation

In response to Vijay's note about extension names[1], I've renamed the example
extensions (correcting a CBOR mismatch in the process). I also noted in section
7.1, Extension identifiers, that  "Use of dot-separated notation here does not
imply an object hierarchy.", and provided a counterexample of a differently-
separated, versioned id: `mycompany- myextension_v01`. None of the pre-defined
extensions were modified.

1) https://lists.w3.org/Archives/Public/public-webauthn/2016Jul/0046.html

* Review update: Use CamelCase for extension identifiers

* Use underscore delimiters in extension IDs

Per the phone call on 13 July 2016, change extension IDs to a format of being
{defining body}_{extension name}<_{version}>.