Fix #577 - CDDL for attStmtTemplate is ambiguous
There are multiple definitions of `attStmtType`; the template defines it
to be `bytes`, while each concrete instance of the template defines it
as a map. This clarifies that it is always a map, since the ".within" control
operator for CDDL defines that the socket `$$attStmtType` to be the superset of
`attStmtTemplate`. [1]

[1] https://tools.ietf.org/html/draft-ietf-cbor-cddl-00#section-3.8.5