Fix #618 - Make PublicKeyCredential.isPlatformAuthenticatorAvailable static
The example in the spec shows static usage of isPlatformAuthenticatorAvailable,
but isPlatformAuthenticatorAvailable is defined [Unscopable], which is a regular
operation which is defined as being not static.

I think the static method of using this is what we want, actually. Declaring
this as a non-static method on PublicKeyCredential means users need to obtain a
concrete PublicKeyCredential object on which to call
isPlatformAuthenticatorAvailable(). This leads to a situation where you have to
first complete a call to navigator.credentials.create() in order to call
isPlatformAuthenticatorAvailable().