Fix #574 - Change language for Create and Get to support hotplugging
This is an incomplete fix; a full fix is intended to be handled in Issue #613.

This reorders the Create and Get operations to indicate that the algorithms for
interacting with devices should be applied as devices are hotplugged / arrive.
It does not specify what happens when devices are removed, nor does it use
precise language. I'm not sure what language would be appropriate in this world,
so this patch is just to make things "better" not "correct".

Resolve @equalsJeffH's comments:

1. Define |lifetimeTimer| and make it available to the line that starts the
   hotplugging
2. Use the |lifetimeTimer| for references later in those algorithms to reduce
   confusion
3. Reword the Notes
4. Change the Notes to Issues