build on Adding a choice for RP to express preferences for attestation types (#693)
* Adding a choice for RP to express preferences for attestation types

Adding a choice for RP to express preferences for attestation types. "low-cost" means don't bother the user to get the attestation, don't do a network roundtrip, etc., even if that means that we won't end up with a meaningful attestation (i.e., even if we only get self-attestation). "verifiable" means that the RP wants a verifiable attestation, even if that means that the user agent needs to connect to the Privacy CA, collect user consent, wait for ECDAA signatures to finish, etc.

* Update index.bs

* Update index.bs

* Added a third option

And changed "low-cost" to "none".

* adding client processing rules...

...for the attestation preferences.

* typo

* typos

* PR 636: rename several items and remove prescriptive wording about UI.

* Addressing akshayku's comments on previous change.

* attestationPresentationPreference -> attestationConveyancePreference

* refine/polich switch stmt

* more refining

* refining AttestationConveyancePreference section

* polish AttestationConveyancePreference section, plus make BS happy

* add issue

* make BS happy

* polish

* need to define 'blinding'

* editorial polish

* fix up MakePublicKeyCredentialOptions.attestation description, remove line-ending blanks

* make BS happy - no comments in <pre> blocks, define attestation conveyance

* adapt commit 825cce7 language

* move attestation conveyance \<dfn>

* address jyasskin comments, thx!