Emphasise that the CollectedClientData can be extended.
There is a risk that RPs will implement overly simplistic parsing of the
CollectedClientData and end up intolerant of any future additions. This
change emphasises the need to parse it properly and will be coupled with
a behaviour in Chrome that inserts a key (`new_keys_may_be_added_here`)
in a random 20% of cases.