Fix #1268 Type transport strings as DOMStrings. (#1275)
* Type transport strings as DOMStrings.

Currently transports are represented as an enum. However, WebIDL has
strict enums. (I.e. an RP which sent an unrecognised transport would
make the whole structure unparsable.) This means that every time we add
a transport, we break all existing browsers.

In order to address this, this change retypes transports as plain
DOMStrings. The AuthenticatorTransport enum still exists, but now only
as documentation and registry — not as a factor in type-checking.

* Address Emil's comment

Fixes #1268.

* Address Jeff's comments.