Create a way to reference the following figure (#1323)
* Address Jeff and Emil's comments.

* Create a way to reference the following figure.

The fact that we have to keep the figure numbers in sync manually
annoyed me. Bikeshed isn't helping out here so I don't see a way to
avoid this for long-range links, but this change eliminates the need to
manually maintain some of the absolute numbers by providing a way to
insert the number of the following figure when the reference and figure
are close by.