Update .spec-data .bikeshed-include cache
Ran the following set of commands:

./update-bikeshed-cache.sh && \
  git add .spec-data .bikeshed-include && \
  git commit .

It’s necessary either that one of the spec editors run those same
commands periodically, or else the .travis.yml CI build file needs to be
changed to stop using cached files in .spec-data and .bikeshed-include.

Prior to running the above commands and committing the changes, the
cached files in .spec-data and .bikeshed-include were more than 2 years
out of date.