Update description of residentKey.
This change emphasises |residentKey| over |requireResidentKey| and
updates some wording to note that the naming is historic. It adds a note
to highlight one implication of the fact that RPs cannot always know
whether a created credential is discoverable or not. Lastly it adds a
remark in the credProps extension about why the |rk| field may be often
omitted by user agents.

Fixes #1459, #1457.