define "self-signed basic attestation type" "SSBasic"
some authenticators in the wild return, on registration, an attestation certificate that is signed by the private key associated with the public key in the cert. Thus, the attestation certificate is essentially the "(CA) root certificate".

this change defines this sort of attestation as "self-signed basic attestation", "SSBasic" for short.