Fix tokenBinding definition type
Parent commit was generating the error:

    FATAL ERROR: Couldn't find target dict-member 'tokenBinding':
    <span data-dict-member-info="" for="CollectedClientData/tokenBinding"></span>

This is because the surrounding `<div dfn-type="dict-member"
dfn-for="CollectedClientData">` defines the definition type as `dict-member`,
but no such member is defined in the `CollectedClientData` dictionary.

This change overrides the definition type to `dfn`, which is described in the
Bikeshed docs [1] as "for general terms and phrases, and a catch-all for
anything else".

[1]: https://tabatkins.github.io/bikeshed/#dfn-types