Have the DPK sign over everything.
Signing over too little is a common problem in protocols and the DPK
wasn't signing over very much. The problem is that the signature was
within the extension itself and thus couldn't easily cover the
authenticator data.

This change puts the DPK signature next to the normal signature in an
assertion, transforming that field into a CBOR array. That's fine for
assertions, but registration doesn't have such a signature output. Thus
this change drops exercising the DPK during registration:

Firstly, this is in line with the user credential which doesn't sign
during registration unless "self" attestation is used. (And "self"
attestation can be use with the DPK if desired.)

Secondly, adding an extra signature output for registration is awkward.
Putting it in the user credential attestation statement is awkward and
conflicts when user-agents replace attestation statements. Adding a new
CTAP field is possible but seems excessive since, as noted, there has
never been a signature by the user credential.