Add "hybrid" as a transport.
FIDO has chosen the name Hybrid for the transport formerly known as
caBLE. This change adds it as a known transport in WebAuthn. Since the
`AuthenticatorTransport` enum is illustrative and not actually
referenced anywhere in the IDL, this is technically a non-normative
change. Also, RPs only store and regurgitate these strings so no change
in RP processing is needed. Still, it's useful to have these strings
documented.