Don't be so strict about uv with the PRF extension.
Authenticators may have different PRFs for the UV and non-UV case. Thus
setting uv=preferred during an assertion is fraught: it doesn't fully
specify which PRF to use.

However, while implementing this, I ended up feeling that the
prohibition on using uv=preferred was too strong. Sites may reasonably
want to use uv=preferred and to take advantage of available PRF results.
If the evaluation points are global then this isn't so silly as to
justify a prohibition, I suspect.