Specify what an unknown type credential descriptor being ignored means
The spec describes that client platforms MUST ignore any
PublicKeyCredentialDescriptor with an unknown type. However, there is no
further specification about the case when this results in an empty
allowCredentials. It must not be treated as an empty list.

The client MUST return an error if none of the listed credentials can be used
in allowCredentials. For instance, if all of the listed credentials have
PublicKeyCredentialDescriptor with an unknown type, the client MUST throw
NotAllowedError.

Fixes #1748