Fix references to reserved tokenBinding member
Commit 5f45d34abe3e6826f1e676c35a3dc5d349bec2ad in PR #1661 changed
the definition type of `CollectedClientData/tokenBinding` from
"dict-member" to "dfn" to resolve Bikeshed errors about the member not
existing in the `CollectedClientData` definition.

However, autolinks to the reserved `tokenBinding` member still use the
`{{foo}}` syntax, which is only for IDL types, so the autolinks no
longer resolve. This results in Bikeshed instead pulling these
autolinks as external refs from the spec database instead of internal
refs, which results in an entry being added to the "terms defined
elsewhere" index: "[WEBAUTHN-3] defines the following terms:
tokenBinding". The changed definition type also changes the generated
link anchor from "#dom-collectedclientdata-tokenbinding" to
"#collectedclientdata-tokenbinding".

This fixes these issues by declaring the link anchor explicitly, for
backwards compatibility, and by changing the autolinks to use the
`[=foo=]` syntax for autolinks to "dfn"-type definitions.

This also fixes the following Bikeshed lint:

```
LINT: Unexported dfn that's not referenced locally - did you mean to export it?
<dfn data-dfn-type="dfn" data-dfn-for="CollectedClientData" id="collectedclientdata-tokenbinding" data-lt="tokenBinding" data-noexport="by-default">tokenBinding<a href="#collectedclientdata-tokenbinding" class="self-link"></a></dfn>
<dfn data-dfn-type="dfn" data-dfn-for="CollectedClientData" id="collectedclientdata-tokenbinding" data-lt="tokenBinding" data-noexport="by-default">tokenBinding<a href="#collectedclientdata-tokenbinding" class="self-link"></a></dfn>
```