Fix indentation of step to invoke authenticatorMakeCredential
The current indentation renders this step as a sub-step of "For each credential
descriptor C in pkOptions.excludeCredentials", meaning the operation is to be
invoked once per `excludeCredentials` entry and not at all if
`excludeCredentials` is empty. The indentation appears to have been broken in
commit b44009c0bc24ed76f79c94c4bf6a3d5a111439ae (originally
7acc1d5ccb24306956acdeb31e995e8f7c486353) in PR #1366, caused by a mix of tab
and space characters being replaced with just spaces.