Align RP ID string types
`rp.id` in `PublicKeyCredentialCreationOptions` and `rpId` in
`PublicKeyCredentialRequestOptions` represent the same thing, but with
different types. The WG agreed on the 2024-05-15 call that both should
be `USVString`. Strictly speaking this is a breaking change (changing
a type bound in input (contravariant) position to be more
restrictive), but in practice this shouldn't be able to break any
applications since then those credentials wouldn't have worked in
`get()` anyway.