Add the `topOrigin` field to the limited clientData verification algorithm.
I was asked to to the PR for this issue, without looking at the
submitted PR, in order to avoid IPR issues that would arise from a
change by a non-member.

The `topOrigin` field was added the the CollectedClientData, and the
serialization algorithm, but not the verification algorithm. This PR
addresses that.

Fixes #2102