Fix AuthenticationExtensionsAuthenticatorInputs/Outputs CDDL
According to the CDDL grammar, after a control operator (called `ctlop` in the
ABNF grammar), there can only be a `type2` production:
https://datatracker.ietf.org/doc/html/rfc8610#appendix-B

In a `type2` production, wrapping parentheses can only be used to wrap a `type`
production. `tstr => any` is a `group` production, and needs to be wrapped in
curly braces or brackets.

In other words, from a CDDL grammar perspective, this is an invalid type:
  `foo .within ( tstr => any )`

This is valid:
  `foo .within { tstr => any }`

This update fixes the CDDL type definitions that used the `.within` operator
with an invalid type2.